# openapi.yaml
openapi: 3.0.1
info:
  title: API Gestion Mairie
  description: 'API REST pour la gestion des réservations de biens municipaux, des
    utilisateurs, des groupes et des pièces justificatives. Système de coordination
    pour la mairie.'
  contact:
    name: Équipe Mairie
    email: contact@mairie.fr
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Serveur de développement local
security:
- bearer-jwt: []
...
paths:
  /auth/login:
    post:
      tags:
      - Authentification
      summary: Connexion utilisateur
      description: Authentifie un utilisateur et retourne un token JWT
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Identifiants invalides
        '403':
          description: Compte sous tutelle ou inactif
  /auth/register:
    post:
      tags:
      - Authentification
      summary: Inscription utilisateur
      description: Crée un nouveau compte utilisateur
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Compte créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Données invalides ou email déjà utilisé
...
  /admin/impersonate/{userId}:
    post:
      tags:
      - Administration
      summary: Impersonate un utilisateur
      description: Permet à un admin de se connecter en tant qu'un autre utilisateur
        (ADMIN uniquement)
      operationId: impersonateUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Token d'impersonation généré
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
        '400':
          description: Utilisateur invalide
        '403':
          description: Accès refusé
      security:
      - bearer-jwt: []
  /admin/audit-logs:
    get:
      tags:
      - Administration
      summary: Consulte les logs d'audit
      description: Récupère les logs d'audit avec filtres optionnels (ADMIN uniquement)
      operationId: getAuditLogs
      parameters:
      - name: actorId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: targetUserId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: action
        in: query
        required: false
        schema:
          type: string
      - name: dateFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: dateTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: Logs d'audit récupérés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAuditLogDTO'
        '403':
          description: Accès refusé
      security:
      - bearer-jwt: []
components:
  schemas:
    AuthRequest:
      type: object
      properties:
        email:
          type: string
        mot_de_passe:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserDTO'
    RegisterRequest:
      type: object
      properties:
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        mot_de_passe:
          type: string
        telephone:
          type: string
    UserDTO:
      type: object
      properties:
        id_user:
          type: string
          format: uuid
        nom:
          type: string
        prenom:
          type: string
        email:
          type: string
        telephone:
          type: string
        role:
          type: string
          enum:
          - ADMIN
          - USER
          - MODERATEUR
          - TUTORED
          - SECRETARY
        statut:
          type: string
          enum:
          - ACTIF
          - INACTIF
          - SUSPENDU
          - PENDING
          - ACTIVE
          - REJECTED
...
    AuditLogDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        actor_role:
          type: string
          enum:
          - ADMIN
          - USER
          - MODERATEUR
          - TUTORED
          - SECRETARY
        target_user_id:
          type: string
          format: uuid
        action:
          type: string
        payload:
          type: string
        timestamp:
          type: string
          format: date-time
...
    PageAuditLogDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
  securitySchemes:
    bearer-jwt:
      type: http
      description: 'Authentification JWT. Format: Bearer {token}'
      scheme: bearer
      bearerFormat: JWT